feat: use cwd (#16)
issues:
* feat: use cwd (#15)

commits:
* feat: use cwd (d7c4def)
