+++ package.json
@@ -1,10 +1,9 @@
 {
   "name": "@mnrendra/obtain-tsconfig-paths",
   "version": "1.1.1",
-  "description": "Obtain a valid baseUrl and paths from the compilerOptions in tsconfig.json file.",
+  "description": "Obtain a valid baseUrl and paths from the compilerOptions in the tsconfig.json file",
   "private": false,
   "main": "dist/index.js",
-  "module": "dist/index.mjs",
   "types": "dist/index.d.ts",
   "scripts": {
     "build": "rm -rf dist && npx rollup -c",
@@ -26,29 +25,29 @@
   },
   "homepage": "https://github.com/mnrendra/obtain-tsconfig-paths#readme",
   "dependencies": {
-    "@mnrendra/read-tsconfig": "^1.2.0"
+    "@mnrendra/read-tsconfig": "^1.3.0"
   },
   "devDependencies": {
     "@semantic-release/exec": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
-    "@types/jest": "^29.5.12",
-    "@types/node": "^20.14.9",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.9.0",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "clean-package": "^2.2.0",
     "cz-conventional-changelog": "^3.3.0",
-    "eslint": "^8.57.0",
+    "eslint": "^8.57.1",
     "eslint-config-standard-with-typescript": "^43.0.1",
-    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-n": "^16.6.2",
-    "eslint-plugin-promise": "^6.4.0",
+    "eslint-plugin-promise": "^6.6.0",
     "jest": "^29.7.0",
-    "rollup": "^4.18.0",
+    "rollup": "^4.26.0",
     "rollup-plugin-dts": "^6.1.1",
     "rollup-plugin-esbuild": "^6.1.1",
-    "semantic-release": "^24.0.0",
-    "ts-jest": "^29.1.5",
+    "semantic-release": "^24.2.0",
+    "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
-    "typescript": "^5.5.3"
+    "typescript": "^5.6.3"
   },
   "keywords": [
     "obtain-tsconfig-paths",
