+++ src/index.ts
@@ -1,24 +1,18 @@
-import { validateSkippedStacks } from '@mnrendra/read-stacked-json'
-
 import async from './async'
 import sync from './sync'
 
 export type {
-  // @mnrendra/types-tsconfig
   TSConfig,
   CompilerOptions,
   BaseURL,
-  Paths,
-  // @mnrendra/obtain-tsconfig-paths
-  Options,
-  TSConfigPaths,
-  // @mnrendra/validate-skipped-stacks
-  SkippedStacks,
-  ValidSkippedStacks
+  Paths
+} from '@mnrendra/read-tsconfig'
+
+export type {
+  TSConfigPaths
 } from './types'
 
 export {
   async as obtainTSConfigPaths,
-  sync as obtainTSConfigPathsSync,
-  validateSkippedStacks
+  sync as obtainTSConfigPathsSync
 }
