+++ src/types/index.ts
@@ -1,16 +1,5 @@
-import type Options from './Options'
 import type TSConfigPaths from './TSConfigPaths'
 
 export type {
-  TSConfig,
-  CompilerOptions,
-  BaseURL,
-  Paths,
-  SkippedStacks,
-  ValidSkippedStacks
-} from '@mnrendra/read-tsconfig'
-
-export type {
-  Options,
   TSConfigPaths
 }
