+++ tests/dummies/index.ts
@@ -1,7 +1,5 @@
 import tsConfigPaths from './tsConfigPaths'
-import tsConfigValues from './tsConfigValues'
 
 export {
-  tsConfigPaths,
-  tsConfigValues
+  tsConfigPaths
 }
