+++ tests/mocks/readTSConfig.ts
@@ -1,0 +1,5 @@
+import * as originalModule from '@mnrendra/read-tsconfig'
+
+const { readTSConfig } = originalModule as jest.Mocked<typeof originalModule>
+
+export default readTSConfig
